07-Data-Engineering/04-Predict-in-production/03-GCR-cloud-run/requirements.txt
gcsfs
pandas
scikit-learn
google-cloud-storage
mlflow
joblib
numpy
psutil
pygeohash
termcolor
xgboost
memoized-property
scipy
category_encoders
s3fs
fastapi
pydantic
uvicorn
